Battle components updated
