+++ src/components/Battle/BattlePage/BattlePage.module.scss
