+++ src/components/Battle/BattlePreview/BattlePreview.module.scss
