+++ src/components/Run/MoveCard/MoveCard.tsx
@@ -9,17 +9,6 @@ type Props = {
 };
 
 const MoveCard: React.FC<Props> = (props: Props) => {
-    // Compute the power of a move
-    const getMovePower = (): string | number => {
-        if (props.move.power === 0) {
-            return "--";
-        } else if (props.isSTAB) {
-            return props.move.power * 1.5;
-        } else {
-            return props.move.power;
-        }
-    };
-
     return (
         <div className={styles["move-card"]}>
             <div className={styles.row}>
@@ -38,12 +27,9 @@ const MoveCard: React.FC<Props> = (props: Props) => {
                         />
                     </div>
                 </div>
-                <p className={styles.text}>
-                    BP:{" "}
-                    <strong className={props.isSTAB && props.move.power ? styles["stab-text"] : ""}>
-                        {getMovePower()}
-                    </strong>
-                    {props.isSTAB && props.move.power > 0 ? (
+                {props.isSTAB && props.move.power > 0 ? (
+                    <p className={styles.text}>
+                        BP: <strong className={styles["stab-text"]}>{props.move.power * 1.5}</strong>
                         <div className={styles["stab-icon"]}>
                             <Image
                                 src="/assets/icons/double-arrow.svg"
@@ -52,10 +38,12 @@ const MoveCard: React.FC<Props> = (props: Props) => {
                                 objectFit="contain"
                             />
                         </div>
-                    ) : (
-                        ""
-                    )}
-                </p>
+                    </p>
+                ) : (
+                    <p className={styles.text}>
+                        BP: <strong>{props.move.power === 0 ? "--" : props.move.power}</strong>
+                    </p>
+                )}
                 <p className={styles.text}>
                     PP: <strong>{props.move.pp}</strong>
                 </p>
