+++ src/models/MoveData.ts
@@ -1,4 +1,5 @@
 export default interface MoveData {
+    slug: string;
     name: string;
     type: string;
     power: number;
